Move WorkbenchAccessTestTrait to its own folder.
