+++ tests/src/Functional/NodeFormTest.php
@@ -5,6 +5,7 @@ namespace Drupal\Tests\workbench_access\Functional;
 use Drupal\taxonomy\Entity\Term;
 use Drupal\Tests\BrowserTestBase;
 use Drupal\workbench_access\WorkbenchAccessManagerInterface;
+use Drupal\Tests\workbench_access\Traits\WorkbenchAccessTestTrait;
 
 /**
  * Tests for the node form.
