+++ tests/src/Functional/TaxonomySchemeUiTest.php
@@ -4,6 +4,7 @@ namespace Drupal\Tests\workbench_access\Functional;
 
 use Drupal\Tests\BrowserTestBase;
 use Drupal\workbench_access\Entity\AccessSchemeInterface;
+use Drupal\Tests\workbench_access\Traits\WorkbenchAccessTestTrait;
 
 /**
  * Defines a class for testing the UI to create and configure schemes.
